+++ CHANGELOG.md
@@ -9,7 +9,9 @@ All notable changes to this project will be documented in this file. For commit 
 
  **BugFixes**:
  - fixed onlyoffice bug https://github.com/gtsteffaniak/filebrowser/issues/418
- - files in breadcrumbs issue https://github.com/gtsteffaniak/filebrowser/issues/419
+ - fixed breadcrumbs bug https://github.com/gtsteffaniak/filebrowser/issues/419
+ - fixed search context bug https://github.com/gtsteffaniak/filebrowser/issues/417
+ - fixed sessionID for search
 
 ## v0.6.0-beta